+++ src/api/PdfMaker.ts
@@ -1,5 +1,5 @@
 import { FontStore } from '../font-store.ts';
-import { ImageStore } from '../image-store.ts';
+import { createImageLoader } from '../image-loader.ts';
 import { layoutPages } from '../layout/layout.ts';
 import type { MakerCtx } from '../maker-ctx.ts';
 import { readDocumentDefinition } from '../read-document.ts';
@@ -18,12 +18,11 @@ export type FontConfig = {
  * Generates PDF documents.
  */
 export class PdfMaker {
-  #ctx: MakerCtx;
+  #fontStore: FontStore;
+  #resourceRoot?: string;
 
   constructor() {
-    const fontStore = new FontStore();
-    const imageStore = new ImageStore();
-    this.#ctx = { fontStore, imageStore };
+    this.#fontStore = new FontStore();
   }
 
   /**
@@ -35,7 +34,7 @@ export class PdfMaker {
    * the meta data cannot be extracted from the font.
    */
   registerFont(data: Uint8Array, config?: FontConfig): void {
-    this.#ctx.fontStore.registerFont(data, config);
+    this.#fontStore.registerFont(data, config);
   }
 
   /**
@@ -45,7 +44,7 @@ export class PdfMaker {
    * @param root The root directory to read resources from.
    */
   setResourceRoot(root: string): void {
-    this.#ctx.imageStore.setResourceRoot(root);
+    this.#resourceRoot = root;
   }
 
   /**
@@ -56,19 +55,14 @@ export class PdfMaker {
    */
   async makePdf(definition: DocumentDefinition): Promise<Uint8Array> {
     const def = readAs(definition, 'definition', readDocumentDefinition);
-    const ctx = { ...this.#ctx };
+    let fontStore = this.#fontStore;
     if (def.fonts) {
-      ctx.fontStore = new FontStore(def.fonts);
+      fontStore = new FontStore(def.fonts);
       console.warn(
         'Registering fonts via document definition is deprecated. Use PdfMaker.registerFont() instead.',
       );
     }
-    if (def.images) {
-      ctx.imageStore = new ImageStore(def.images);
-      console.warn(
-        'Registering images via document definition is deprecated. Use URLs to include images instead.',
-      );
-    }
+    const ctx: MakerCtx = { fontStore, imageLoader: createImageLoader(this.#resourceRoot) };
     if (def.dev?.guides != null) ctx.guides = def.dev.guides;
     const pages = await layoutPages(def, ctx);
     return await renderDocument(def, pages);