+++ src/maker-ctx.ts
@@ -1,8 +1,8 @@
 import type { FontStore } from './font-store.ts';
-import type { ImageStore } from './image-store.ts';
+import type { ImageLoader } from './image-loader.ts';
 
 export type MakerCtx = {
   fontStore: FontStore;
-  imageStore: ImageStore;
+  imageLoader: ImageLoader;
   guides?: boolean;
 };